test suite for responses API
